requirements.txt
google-cloud-storage
Flask
tensorflow
numpy
